Refactor: update name of ServerStatusInterface
