+++ src/main/java/ServerStatus.java
@@ -1,8 +1,3 @@
-public class ServerStatus implements ServerStatusInterface {
-
-    private boolean keepAcceptingConnections = true;
-
-    public boolean acceptConnections() {
-        return keepAcceptingConnections;
-    }
+public interface ServerStatus {
+    boolean acceptConnections();
 }
