--- src/main/java/ServerStatusInterface.java
@@ -1,3 +1,0 @@
-public interface ServerStatusInterface {
-    boolean acceptConnections();
-}
