+++ src/main/java/WebServerStatus.java
@@ -1,0 +1,8 @@
+public class WebServerStatus implements ServerStatus {
+
+    private boolean keepAcceptingConnections = true;
+
+    public boolean acceptConnections() {
+        return keepAcceptingConnections;
+    }
+}
