+++ src/test/java/WebServerStatusTest.java
@@ -1,0 +1,10 @@
+import org.junit.Assert;
+import org.junit.Test;
+
+public class WebServerStatusTest {
+    @Test
+    public void statusAlwaysListening() {
+        WebServerStatus webServerStatus = new WebServerStatus();
+        Assert.assertEquals(true, webServerStatus.acceptConnections());
+    }
+}
